third_party/WebKit/LayoutTests/gamepad/gamepad-detached-no-crash.html

<!DOCTYPE HTML>
<script src="../resources/js-test.js"></script>
<script>
description("Accessing Navigator Gamepad methods on a closed window.");

window.jsTestIsAsync = true;

var expectedLength = 0;
var w;
function processMessage(event) {
    if (event.data == "opened") {
        w.close();
    } else if (event.data == "closed") {
        shouldBe("w.navigator.getGamepads().length", "expectedLength");
        finishJSTest();
    }
}

if (window.gamepadController) {
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.setCanOpenWindows();
        testRunner.waitUntilDone();
    }
    w = window.open('../resources/window-postmessage-open-close.html');
    expectedLength = w.navigator.getGamepads().length;
    window.addEventListener("message", processMessage, false);
} else {
    console.log("FAIL: no gamepad controller available.")
}
</script>
